Add OpenConsole support for Windows
If the conpty.dll and OpenConsole.exe files from the Windows Terminal
project are placed in the same directory as the alacritty executable
they will be used instead of implementation provided by Windows.

This way you don't need to wait for new Windows releases to get bug
fixes and new features.